+++ RELEASE_NOTES.md
@@ -7,6 +7,8 @@ _Date TBD_
 * Added Google Tag Manager event tracking
 * Added POC to Product level; refactored Contact as separate component
 * Allow splitting of Products along Version lines
+* Show retired/deleted ONC-ACBs on main page search filter
+* Allow optionally banning Developers on appropriate status change
 
 ---